requirements.txt
fastapi
uvicorn
pillow
tensorflow==2.13.0
numpy
python-multipart
